+++ tests/nativeapp/test_init.py
@@ -1,7 +1,6 @@
 from snowcli.cli.nativeapp.init import (
     render_snowflake_yml,
     _sparse_checkout,
-    _move_and_rename_project,
     render_nativeapp_readme,
     nativeapp_init,
     CannotInitializeAnExistingProjectError,
@@ -26,24 +25,6 @@ def test_sparse_checkout(other_directory):
     assert cloned_repo.exists()
     assert cloned_repo.joinpath("samples").exists()
     assert not cloned_repo.joinpath("README.md").exists()
-
-
-def test_move_and_rename_project(other_directory):
-    temp_dir = other_directory
-    temp_dir_path = Path(temp_dir)
-    directory_to_make = temp_dir_path.joinpath("some/random/path/samples")
-    directory_to_make.mkdir(parents=True, exist_ok=False)
-    _move_and_rename_project(
-        source_parent_directory=directory_to_make.parent,
-        target_parent_directory=temp_dir_path,
-        repo_sub_directory="samples",
-        new_name="moved_samples_directory",
-    )
-
-    assert temp_dir_path.exists()
-    assert temp_dir_path.joinpath("moved_samples_directory").exists()
-    assert directory_to_make.parent.exists()
-    assert not directory_to_make.exists()
 
 
 def test_render_snowflake_yml(other_directory):
